fix: compile error

+++ src/Wolfe.SpaceTraders.Sdk/Models/Contracts/SpaceTradersContractPaymentTerms.cs
@@ -2,6 +2,6 @@
 
 public class SpaceTradersContractPaymentTerms
 {
-    public int OnAccepted { get; set; }
-    public int OnFulfilled { get; set; }
+    public long OnAccepted { get; set; }
+    public long OnFulfilled { get; set; }
 }
